Create vote.php
Test page for the voting system for clearskyy. Users are able to vote on reviews that the site has and the vote are stored in a vote database. Right now its totally busted. It takes the last URL pulled from the loop and takes the previous vote value, so if you click the first time, nothing is passed to the post, but if you click another vote the previous vote is passed. There is nothing in place yet to limit votes, but it will be limited by IP down the road.